Added V, modified dt formula, modified bounce back velocities condition
